requirements.txt
absl-py==0.8.1
aiohttp==3.6.2
astor==0.8.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
cachetools==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cssselect2==0.2.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
digi-xbee==1.3.0
dlib==19.18.0
docutils==0.15.2
dronekit==2.9.2
easytello==0.0.7
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
future==0.18.2
gast==0.2.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
h5py==2.10.0
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.15.0
idna==2.8
idna-ssl==1.1.0
imageio==2.6.1
importlib-metadata==0.23
imutils==0.5.3
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.3
joblib==0.14.0
jsmin==2.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.3.0
kiwisolver==1.1.0
livereload==2.6.1
loguru==0.3.2
lxml==4.4.2
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkdocs==1.0.4
mkdocs-material==4.5.0
mkdocs-minify-plugin==0.2.1
mkdocs-pdf-export-plugin==0.5.5
monotonic==1.5
more-itertools==7.2.0
mss==4.0.3
mtcnn==0.1.0
multidict==4.7.3
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.1
pafy==0.5.5
pandas==0.25.3
pandocfilters==1.4.2
parameterized==0.7.1
parso==0.5.1
pep562==1.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pydot==1.4.1
PyDrive==1.3.1
Pygments==2.4.2
pygpu==0.7.6
pymavlink==2.4.2
pymdown-extensions==6.2
pyparsing==2.4.5
Pyphen==0.9.5
pyrsistent==0.15.6
pyserial==3.4
pyswip==0.2.9
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzbar==0.1.8
pyzmq==18.1.1
qrcode==6.1
qrcode-terminal==0.8
qtconsole==4.6.0
readme-renderer==24.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-fuzzy==0.4.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.3
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
srp==1.0.15
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
Theano==1.0.4
tinycss2==1.0.2
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
tta-wrapper==0.0.1
twine==3.1.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.7
vidgear==0.1.5
vidstab==1.7.2
wcwidth==0.1.7
WeasyPrint==50
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
XBee==2.3.2
xlrd==1.2.0
xlwt==1.3.0
yarl==1.4.2
youtube-dl==2019.11.28
zipp==0.6.0
